+++ includes/system/class-statistics.php
@@ -6,10 +6,10 @@
  *
  * @package System
  * @author  Pierre Lannoy <https://pierre.lannoy.fr/>.
- * @since   1.0.0
+ * @since   1.3.0
  */
 
-namespace WPPluginBoilerplate\System;
+namespace Decalog\System;
 
 /**
  * Define the plugin statistics functionality.
@@ -18,22 +18,30 @@ namespace WPPluginBoilerplate\System;
  *
  * @package System
  * @author  Pierre Lannoy <https://pierre.lannoy.fr/>.
- * @since   1.0.0
+ * @since   1.3.0
  */
 class Statistics {
 
 	/**
+	 * The unique instance of the class.
+	 *
+	 * @since  1.3.0
+	 * @var self $instance The unique instance of the class.
+	 */
+	private static $instance;
+
+	/**
 	 * The number of active installs for this plugin.
 	 *
-	 * @since  1.0.0
-	 * @var    integer    $active_installs    The number of active installs.
+	 * @since  1.3.0
+	 * @var    integer    $installs    The number of active installs.
 	 */
-	public $active_installs = -1;
+	public $installs = -1;
 
 	/**
 	 * The number of active downloads for this plugin.
 	 *
-	 * @since  1.0.0
+	 * @since  1.3.0
 	 * @var    integer    $downloads    The number of downloads.
 	 */
 	public $downloads = -1;
@@ -41,7 +49,7 @@ class Statistics {
 	/**
 	 * The rating of this plugin.
 	 *
-	 * @since  1.0.0
+	 * @since  1.3.0
 	 * @var    integer    $rating    The rating.
 	 */
 	public $rating = -1;
@@ -49,7 +57,7 @@ class Statistics {
 	/**
 	 * The number of reviews for this plugin.
 	 *
-	 * @since  1.0.0
+	 * @since  1.3.0
 	 * @var    integer    $reviews    The number of reviews.
 	 */
 	public $reviews = -1;
@@ -57,7 +65,7 @@ class Statistics {
 	/**
 	 * Initializes the class and set its properties.
 	 *
-	 * @since 1.0.0
+	 * @since 1.3.0
 	 */
 	public function __construct() {
 		$this->get_wp_stats();
@@ -66,7 +74,7 @@ class Statistics {
 	/**
 	 * Get the WP stats about active installs, downloads, rating and reviews.
 	 *
-	 * @since   1.0.0
+	 * @since   1.3.0
 	 */
 	private function get_wp_stats() {
 		$stats = Cache::get_global( 'self_wp_stats' );
@@ -84,7 +92,7 @@ class Statistics {
 				$api   = plugins_api(
 					'plugin_information',
 					[
-						'slug'   => WPPB_SLUG,
+						'slug'   => DECALOG_SLUG,
 						'fields' => $query,
 					]
 				);
@@ -101,7 +109,7 @@ class Statistics {
 		}
 		if ( false !== $stats ) {
 			if ( array_key_exists( 'active_installs', $stats ) ) {
-				$this->active_installs = $stats['active_installs'];
+				$this->installs = $stats['active_installs'];
 			}
 			if ( array_key_exists( 'downloaded', $stats ) ) {
 				$this->downloads = $stats['downloaded'];
@@ -113,5 +121,41 @@ class Statistics {
 				$this->reviews = $stats['num_ratings'];
 			}
 		}
+	}
+
+	/**
+	 * Get the statistics as shortcode.
+	 *
+	 * @param   array $attributes  Attributes of the shortcode.
+	 * @return  int|string  The output of the shortcode, ready to print.
+	 * @since   1.3.0
+	 */
+	public static function sc_get_raw( $attributes ) {
+		$_attributes = shortcode_atts(
+			[
+				'item' => 'rating',
+			],
+			$attributes
+		);
+		if ( ! isset( self::$instance ) ) {
+			self::$instance = new Statistics();
+		}
+		switch ( $_attributes['item'] ) {
+			case 'installs':
+				$result = self::$instance->installs;
+				break;
+			case 'downloads':
+				$result = self::$instance->downloads;
+				break;
+			case 'rating':
+				$result = self::$instance->rating;
+				break;
+			case 'reviews':
+				$result = self::$instance->reviews;
+				break;
+			default:
+				$result = '';
+		}
+		return $result;
 	}
 }