+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to **DecaLog** is documented in this *changelog*.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and **DecaLog** adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- New shortcodes to query the plugin statistics.
 
 ## [1.2.1] - 2019-09-05
 ### Changed